Refactor device views into directory Refactors internals to named exports Updated per Ani's review
OKTA-374132
<<<Jenkins Check-In of Tested SHA: bdc45118718f887180017f888d5add8686b1bf0d for eng_productivity_ci_bot_okta@okta.com>>>
Artifact: okta-signin-widget
