+++ src/v2/view-builder/views/ov/ChallengeAuthenticatorDataOktaVerifyView.js
@@ -1,5 +1,5 @@
 import { Collection } from 'okta';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm } from '../../internals';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import AuthenticatorVerifyOptions from '../../components/AuthenticatorVerifyOptions';
 import { getAuthenticatorDataForVerification } from '../../utils/AuthenticatorUtil';
