+++ src/v2/view-builder/views/phone/ChallengeAuthenticatorDataPhoneView.js
@@ -1,6 +1,5 @@
 import { _, loc } from 'okta';
-import BaseView from '../../internals/BaseView';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm, BaseView } from '../../internals';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import AuthenticatorVerifyFooter from '../../components/AuthenticatorVerifyFooter';
